Use @hapi/hapi and refactor index into server file
