+++ index.js
@@ -1,26 +1,5 @@
 'use strict';
 
-const Hapi = require('hapi');
-var models = require('./models');
+const { start } = require('./lib/server');
 
-const init = async () => {
-
-    const server = Hapi.server({
-        port: 3000,
-        host: 'localhost'
-    });
-
-    server.route(require('./lib/routes'));
-
-    server.start(function() {
-      console.log('Running on 3000');
-    });
-};
-
-process.on('unhandledRejection', (err) => {
-
-    console.log(err);
-    process.exit(1);
-});
-
-init();
+start();
