+++ package.json
@@ -18,10 +18,11 @@
   },
   "homepage": "https://github.com/Tweet-311-Denver/tweet_311_denver_service#readme",
   "dependencies": {
-    "puppeteer": "^2.1.1",
+    "@hapi/hapi": "^19.1.1",
     "hapi": "^18.1.0",
     "pg": "^7.18.1",
     "pg-hstore": "^2.3.3",
+    "puppeteer": "^2.1.1",
     "sequelize": "^5.21.4",
     "sequelize-cli": "^5.5.1"
   }
